refactor(brpsystems): move logic to /providers
