+++ rezervo/providers/brpsystems/booking.py
@@ -7,9 +7,9 @@ import requests
 
 from rezervo.consts import WEEKDAYS
 from rezervo.errors import AuthenticationError, BookingError
-from rezervo.integrations.ttt.auth import authenticate
-from rezervo.integrations.ttt.schedule import fetch_fsc_schedule
-from rezervo.integrations.ttt.schema import (
+from rezervo.providers.brpsystems.auth import authenticate
+from rezervo.providers.brpsystems.schedule import fetch_fsc_schedule
+from rezervo.providers.brpsystems.schema import (
     BookingData,
     BookingType,
     BrpAuthResult,
